test: accelerate the edit-loop suite
