+++ test/minga_editor/ui/picker/file_source_test.exs
@@ -4,11 +4,9 @@ defmodule MingaEditor.UI.Picker.FileSourceTest do
   # Uses the global Minga.Project singleton to drive FileSource.project_root/0.
   use Minga.Test.EditorCase, async: false
 
-  alias MingaEditor.PickerUI
-  alias MingaEditor.State.ModalOverlay
-  alias MingaEditor.State.ModalOverlay.Picker, as: PickerPayload
-  alias MingaEditor.State.Picker, as: PickerState
-  alias MingaEditor.UI.Picker
+  alias MingaEditor.RenderPipeline.TestHelpers
+  alias MingaEditor.State, as: EditorState
+  alias MingaEditor.State.FileTree
   alias MingaEditor.UI.Picker.Context
   alias MingaEditor.UI.Picker.FileSource
   alias MingaEditor.UI.Picker.Item
@@ -52,135 +50,6 @@ defmodule MingaEditor.UI.Picker.FileSourceTest do
 
     assert Minga.Buffer.file_path(state.workspace.buffers.active) == Path.join(lib, "hot.ex")
     assert Minga.Project.frecency_scores()["lib/hot.ex"] > 0
-  end
-
-  test "on_bulk_select opens all marked project-relative files", %{tmp_dir: tmp_dir} do
-    project = Path.join(tmp_dir, "frecency_bulk_project_#{:erlang.unique_integer([:positive])}")
-    lib = Path.join(project, "lib")
-    File.mkdir_p!(lib)
-    File.write!(Path.join(project, "mix.exs"), "")
-    File.write!(Path.join(project, "initial.ex"), "initial")
-    File.write!(Path.join(lib, "one.ex"), "one")
-    File.write!(Path.join(lib, "two.ex"), "two")
-
-    Minga.Events.subscribe(:project_rebuilt)
-    Minga.Project.switch(project)
-    await_project_rebuild(project)
-
-    ctx =
-      start_editor("initial", file_path: Path.join(project, "initial.ex"), project_root: project)
-
-    state = editor_state(ctx)
-    initial_pids = state.workspace.buffers.list
-
-    state =
-      FileSource.on_bulk_select(
-        [%Item{id: "lib/one.ex", label: "one.ex"}, %Item{id: "lib/two.ex", label: "two.ex"}],
-        state
-      )
-
-    paths = Enum.map(state.workspace.buffers.list, &Minga.Buffer.file_path/1)
-    new_pids = Enum.reject(state.workspace.buffers.list, &Enum.member?(initial_pids, &1))
-
-    on_exit(fn -> stop_pids(new_pids) end)
-
-    assert Path.join(lib, "one.ex") in paths
-    assert Path.join(lib, "two.ex") in paths
-    assert Minga.Buffer.file_path(state.workspace.buffers.active) == Path.join(lib, "two.ex")
-  end
-
-  test "PickerUI Enter opens marked files through the UI path", %{tmp_dir: tmp_dir} do
-    project =
-      Path.join(tmp_dir, "frecency_ui_select_project_#{:erlang.unique_integer([:positive])}")
-
-    lib = Path.join(project, "lib")
-    File.mkdir_p!(lib)
-    File.write!(Path.join(project, "mix.exs"), "")
-    File.write!(Path.join(project, "initial.ex"), "initial")
-    File.write!(Path.join(lib, "one.ex"), "one")
-    File.write!(Path.join(lib, "two.ex"), "two")
-
-    Minga.Events.subscribe(:project_rebuilt)
-    Minga.Project.switch(project)
-    await_project_rebuild(project)
-
-    ctx =
-      start_editor("initial", file_path: Path.join(project, "initial.ex"), project_root: project)
-
-    state = editor_state(ctx)
-    initial_pids = state.workspace.buffers.list
-
-    picker_state =
-      build_file_picker_state(state, [
-        %Item{id: "lib/one.ex", label: "one.ex"},
-        %Item{id: "lib/two.ex", label: "two.ex"}
-      ])
-
-    new_state = PickerUI.handle_key(picker_state, 13, 0)
-    flush_project()
-    paths = Enum.map(new_state.workspace.buffers.list, &Minga.Buffer.file_path/1)
-    new_pids = Enum.reject(new_state.workspace.buffers.list, &Enum.member?(initial_pids, &1))
-
-    on_exit(fn -> stop_pids(new_pids) end)
-
-    assert new_state.shell_state.modal == :none
-    assert Path.join(lib, "one.ex") in paths
-    assert Path.join(lib, "two.ex") in paths
-    assert Minga.Buffer.file_path(new_state.workspace.buffers.active) == Path.join(lib, "two.ex")
-  end
-
-  test "PickerUI bulk action menu dispatches on_bulk_action for marked files", %{tmp_dir: tmp_dir} do
-    project =
-      Path.join(tmp_dir, "frecency_ui_bulk_project_#{:erlang.unique_integer([:positive])}")
-
-    lib = Path.join(project, "lib")
-    File.mkdir_p!(lib)
-    File.write!(Path.join(project, "mix.exs"), "")
-    File.write!(Path.join(project, "initial.ex"), "initial")
-    File.write!(Path.join(lib, "one.ex"), "one")
-    File.write!(Path.join(lib, "two.ex"), "two")
-
-    Minga.Events.subscribe(:project_rebuilt)
-    Minga.Project.switch(project)
-    await_project_rebuild(project)
-
-    ctx =
-      start_editor("initial", file_path: Path.join(project, "initial.ex"), project_root: project)
-
-    state = editor_state(ctx)
-    initial_pids = state.workspace.buffers.list
-
-    picker_state =
-      build_file_picker_state(state, [
-        %Item{id: "lib/one.ex", label: "one.ex"},
-        %Item{id: "lib/two.ex", label: "two.ex"}
-      ])
-
-    menu_state = PickerUI.handle_key(picker_state, ?o, MingaEditor.Input.mod_ctrl())
-
-    assert {:picker,
-            %{picker_ui: %{action_menu: {[{"Open all marked", {:bulk, :open_marked, items}}], 0}}}} =
-             menu_state.shell_state.modal
-
-    assert Enum.map(items, & &1.id) == ["lib/one.ex", "lib/two.ex"]
-
-    new_state = PickerUI.handle_key(menu_state, 13, 0)
-    flush_project()
-    paths = Enum.map(new_state.workspace.buffers.list, &Minga.Buffer.file_path/1)
-    new_pids = Enum.reject(new_state.workspace.buffers.list, &Enum.member?(initial_pids, &1))
-
-    on_exit(fn -> stop_pids(new_pids) end)
-
-    assert new_state.shell_state.modal == :none
-    assert Path.join(lib, "one.ex") in paths
-    assert Path.join(lib, "two.ex") in paths
-    assert Minga.Buffer.file_path(new_state.workspace.buffers.active) == Path.join(lib, "two.ex")
-  end
-
-  test "bulk actions expose open all marked" do
-    assert FileSource.bulk_actions([%Item{id: "lib/one.ex", label: "one.ex"}]) == [
-             {"Open all marked", :open_marked}
-           ]
   end
 
   test "preview selections do not record frecency until confirmed", %{tmp_dir: tmp_dir} do
@@ -257,10 +126,15 @@ defmodule MingaEditor.UI.Picker.FileSourceTest do
     # it; reading the cache must not, proving the picker uses Minga.Project.files/0.
     File.write!(Path.join(project, "added_after_cache.ex"), "late")
 
-    ctx =
-      start_editor("cached", file_path: Path.join(project, "cached.ex"), project_root: project)
+    state =
+      TestHelpers.base_state(content: "cached")
+      |> EditorState.set_file_tree(%FileTree{project_root: project})
 
-    ids = ctx |> editor_state() |> FileSource.candidates() |> Enum.map(& &1.id)
+    ids =
+      state
+      |> Context.from_editor_state()
+      |> FileSource.candidates()
+      |> Enum.map(& &1.id)
 
     assert "cached.ex" in ids
     refute "added_after_cache.ex" in ids
@@ -279,12 +153,11 @@ defmodule MingaEditor.UI.Picker.FileSourceTest do
     Minga.Project.switch(active)
     await_project_rebuild(active)
 
-    ctx = start_editor("active", file_path: Path.join(active, "mix.exs"), project_root: active)
+    state =
+      TestHelpers.base_state(content: "active")
+      |> EditorState.set_file_tree(%FileTree{project_root: active})
 
-    picker_context =
-      ctx
-      |> editor_state()
-      |> Context.from_editor_state(%{project_root: other})
+    picker_context = Context.from_editor_state(state, %{project_root: other})
 
     ids = FileSource.candidates(picker_context) |> Enum.map(& &1.id)
 
@@ -301,71 +174,16 @@ defmodule MingaEditor.UI.Picker.FileSourceTest do
     File.write!(Path.join(stale_root, "stale.txt"), "stale")
     File.write!(Path.join(selected_root, "target.txt"), "target")
 
-    ctx =
-      start_editor("stale",
-        file_path: Path.join(stale_root, "stale.txt"),
-        project_root: stale_root
-      )
+    state =
+      TestHelpers.base_state(content: "stale")
+      |> EditorState.set_file_tree(%FileTree{project_root: stale_root})
 
-    picker_context =
-      ctx
-      |> editor_state()
-      |> Context.from_editor_state(%{project_root: selected_root})
+    picker_context = Context.from_editor_state(state, %{project_root: selected_root})
 
     ids = FileSource.candidates(picker_context) |> Enum.map(& &1.id)
 
     assert "target.txt" in ids
     refute "stale.txt" in ids
-  end
-
-  describe "lean candidates and enrich/1" do
-    test "enrich builds icon, color, two-line description, and git annotation for winners" do
-      lean = %Item{
-        id: "lib/foo/bar.ex",
-        label: "bar.ex",
-        search_text: "lib/foo/bar.ex",
-        meta: %{git: :modified}
-      }
-
-      [enriched] = FileSource.enrich([lean])
-
-      assert enriched.id == "lib/foo/bar.ex"
-      assert String.ends_with?(enriched.label, " bar.ex")
-      assert String.first(enriched.label) != "b"
-      assert enriched.description == "lib/foo"
-      assert enriched.annotation == "M"
-      assert enriched.two_line == false
-      assert is_integer(enriched.icon_color)
-    end
-
-    test "enrich uses an empty description for root-level files and no git annotation" do
-      lean = %Item{id: "mix.exs", label: "mix.exs", search_text: "mix.exs", meta: %{git: nil}}
-
-      [enriched] = FileSource.enrich([lean])
-
-      assert enriched.description == ""
-      assert enriched.annotation == nil
-    end
-  end
-
-  defp build_file_picker_state(state, items) do
-    picker = items |> Picker.new(title: "Find file", max_visible: 10) |> mark_all_picker()
-
-    picker_state = %PickerState{
-      picker: picker,
-      source: FileSource,
-      restore: state.workspace.buffers.active_index
-    }
-
-    ModalOverlay.open(state, :picker, PickerPayload.new(picker_state))
-  end
-
-  defp mark_all_picker(%Picker{items: []} = picker), do: picker
-
-  defp mark_all_picker(%Picker{} = picker) do
-    Enum.reduce(1..length(picker.items), picker, fn _, acc ->
-      Picker.toggle_mark(acc) |> Picker.move_down()
-    end)
   end
 
   defp stop_pids(pids) do
